Using Random Forests and XGBoost
Using a Random Forest and then applying Gradient Boosting techniques, such as XGBoost, to dig for insight in IBM's Employee Attrition data.

Assessing the likelihood that an employee leaves the company based on the features that apply to them. Data has 34 features, 1 target to investigate (Attrition), and 1470 different person's information.

Some insights from this dataset:
Attrition is the highest for both men and women from 18 to 35 years of age and gradually decreases.
As income increases, attrition decreases.
Attrition is much less in divorced women.
Attrition is higher for employees who usually travel more than others, and this rate is higher for women than for men.
Attrition is the highest for those in level 1 jobs.
Women with the job position of manager, research director and technician laboratory have almost no attrition.
Men with the position of sales expert have higher rates of attrition.