Causal Inference - 2SLS Regression with IV
Exercise that studies influencing medical expenses given a endogenous regressor of having health insurance and exogenous regressors of illnesses, age, and income. Instrument variable in the one instrument case is the social security income ratio. Data is from the Medical Expenditure Panel Survey. The results confirm the causal relationship that having health insurance reduces medical expenses.

An endogenous variable (a variable influenced by unobserved factors) affects both the dependent variable and an independent variable, it leads to biased estimates in ordinary least squares (OLS) regression, as the endogenous variable is correlated with the error terms. We pick an IV that is correlated in some way to the endogenous variable and use that as a proxy. Social security income ratio is a IV variable that is correlated with the endogenous explanatory variable but uncorrelated with the error term. It acts as a proxy, capturing the variation in the endogenous variable. IVs break the correlation between the endogenous variable and the error term. In the first step, we regress health insurance on the exogeneous variables and the IV. This regression estimates the causal effect of the endogenous variable on the dependent variable, while accounting for endogeneity using the predicted values from the IV regression. This reduces the endogeneity and we can now more clearly see the impact on medical expenses.

The Durbin-Wu-Hausman test compares OLS and the 2SLS model coefficients. The null hypothesis is that the regressors are exogenous is rejected. Therefore, the health insurance is an endogenous regressor and we need to our instrumental variables approach. There is low correlation among instruments and endogenous variable, of about 0.1-0.25 in absolute value. The correlation is low, but not an indication of a weak instrument. The test for weak instruments looks at the F-statistic for joint significance of instruments. The number is 69 from the model with 1 instruments, which is larger than the rule of thumb of 10. 

Interpretation of the coefficient of the endogenous variable in 2SLS: After instrumentation, for individuals with health insurance, their medical expenses are 85.2% lower compared to individuals without health insurance.