Chatbot using Python with a Streamlit UI_v1
Given the popular applications of Natural Language Processing, I put together a 'version one' of a chatbot in Python. Using the streamlit library in Python I created a simple user interface.

It can handle some general conversation and can provide advice on budgeting and credit scores for now. In future versions I will be training the chatbot with more data to handle fuller conversations, with topic niches in economics and finance, for now this is a basic demo.